Refactor text utilities and fix extraction
